Dev (#5)
* default linting configuration applied, prettier format, and husky pre-commit checks as well

* apply lint staged command

* remove styled lint command from lintstaged rc

* should error out

* ignore eslint cache, and prettier should format the app.tsx file

* configure source path for project

* Configure tailwind css and prepare other assets

Configure tailwindcss to run on this project, configured prettier with
the taliwind rules, have eslint ignore config files, and import assets
for the application

* 0.0.1

* update the node types to properly set the file path alias

* set html structure of the component

set the html structure, put the sign up form into its own compoennt,
apply basic system images

* create sign up logic for component

define a context to be used for the component that contains the target
email address, display errro when invalid email is displayed, confirm
message is displayed and user can navigate back to previous form.

* set tailwind default styles

set base styls for tailwind, customized the config and implemented some
base styles for the mobile view

* set font family

create the custom rule in the `index.css` file and adjsut the tailwind
config to include the font as a ruleset

* style the list of benefits for signing up

* set base style for mobile

edit the tailwind config and create base style for the mobile view of
the newsletter component

* error message

style the error message that appears when a user types in an invalid
email.

* style the confirming page

apply styles for the _confirming email_ component, animate the spinner
and apply new font size to the tailwind config.

* style the confirmation

use tailwindcss to style the confirmation page.

* adjust image

image did not stretch width of screen on larger mobile devices, set
width to full

* fix mistyped state value

* style component for tablet view

apply tailwind breakpoints and classes to style the compoennt for tablet
view.

* structure desktop page

use tailwind css to apply breakpoints and adjust the page structure for
desktop view.

* style sign up for desktop

apply tailwind styles and customize config to style for desktop.

* style desktop view

style the desktop view of the component of the confirmation and
confirmatin pages, adjust tablet view of confirmation page to match
figma design. Tailwind config modified with custom class names.

* update page title and add simple meta description

* ci/cd process

create ci/cd process for the main branch, link project to vercel and
ignore vercel assets in `.gitignore`.

* adjust margin

increase margin for sign up text on lg screens.

* update readme

update the readme with project information, import a gif showcasing the
component in use.

* 1.0.0

* remove cypress action

cypress testing is not used, remove the cypress action from the ci/cd
process.